Create 1527. Patients With a Condition
Can use LIKE or REGEXP
LIKE 'DIAB1%' or LIKE '% DIAB1%'
need to include space if DIAB1 is not the first entry in conditions as expression might start with a different disease

\\b = LIKE 'DIAB1%' or LIKE '% DIAB1%'
"The reason they are the same is that \b matches either a non-word character (in our case, a space) or the position before the first character in the string. Also, you need to escape a backslash with another backslash, like so: \\b. Otherwise, the regular expression won't evaluate."