Create 290. Word Pattern
NEED TO REMEMBER TO CHECK BASE CONDITIONS

Need to verify that the number of tokens in the String match the length of the pattern, plus also need to check if the token is already in the HashMap via containsValue() to check for a collision. (Can also use two hashmaps instead.