Create 61. Rotate List
Count the number of nodes, when you get to the end turn the list into a cycle. Then calculate where the new end of the list should be by calculating #nodes - (k % #nodes). This is where the new break will be because the cycle is broken at difference. then set the new head to the next node at the new break point and set the new breakpoint