Create 3. Longest Substring Without Repeating Characters
Sliding Window Problem.

Two pointers so only have to iterate through the list once. If the letter is not in the HashSet add it and increment the counter. If the letter already exists in the set then the count has to reset. Have to move the trailing pointer up until the point where the repeated character is encountered to "start over".