Create 11. Container With Most Water
Start with pointers at either end (that way the range between values is maximized). Move the pointer which is pointing to the smaller value to the next value and check if the amount of water stored his higher than the previous value. Keep moving the smaller of the pointers and checking the amount of water until the two pointers cross, that way the array only needs to be iterated through 1 time.