Create 841. Keys and Rooms
This solution does not use DFS. Use a hashset to keep track of which rooms have been visited and a queue to keep track of which rooms you have keys for. When you enter a room if one of the keys is to a room has not been visited add that room to the queue. While there are still rooms to visit keep going then check if the hashset is the same size as the number of rooms