Create 45. Jump Game II
Iterate through array keeping track of the farthest index that can be reached by combining the current index with the maximum range. If the current max is larger than the size of the array return the #of hops. If it isnt, increment the number of hops and continue iterating through the array