Create 6. Zigzag Conversion
Have direction bool that switches when reaching the number of rows. Have a separate string builder for each row and as you iterate through the string keep track of the current row and add the current char to each stringbuilder. Direction bool pivots based on the limit of the number of rows.

Strinbuilder has toString() function. Add chars to stringBuilder using .append()

Need special case for row of 1 (just return string)