Create 28. Find the Index of the First Occurrence in a String
Two Pointer problem iterate 1 pointer through haystack. When a matching character occurs in haystack as needle, iterate second pointer through needle comparing each character to the next character in the haystack. When reaching the end of the needle can return the value of needle pointer - size of needle, otherwise return -1 because the needle does not exist in the haystack