Create 128. Longest Consecutive Sequence
O(n) solution with hashset.

Add numbers to HashSet for fast lookup. Every sequence of numbers will have a "start" where the number to the left of the start does not exist. When iterating through the array if the value of nums[i] - 1 is not in the set then the value is a start of a sequence and as long as nums[i] + 1 is in the array keep incrementing the counter and return the largest sequence size (Math.max()) as answer.
Initial attempt is very slow.