Create 209. Minimum Size Subarray Sum
Solution involves sliding window with 2 pointers. Increment the first pointer while the current sum is less than the target increasing the current sum by the amount at the first pointer, then when above or equal to the target subtract and increment the second pointer until the value is less than the target. This allows solution to be found using only 1 pass through the array