Update 111. Minimum Depth of Binary Tree
MY BFS solution using queue. Check if root is null, if it isnt add the root to the queue. Then for each row get the size of the queue (1,2,4,8, etc) Only increment the counter if every node at that depth isnt a leaf node. If any are a leaf node can return the current depth because the minimum depth has been found