Create 153. Find Minimum in Rotated Sorted Array
Sorted array in logn time == Binary search problem. Can check if array is in order by comparing first and last elements. Otherwise keep the left most element saved. If the midpoint of search is less than the current min val, update minval and search to the left of the current midpoint. Otherwise, the minimum of the value must be in the righthandside of the midpoint so update the left value to be the midpoint and search on the right side.