Create 80. Remove Duplicates from Sorted Array II
Two pointer solution. First 2 elements will always be maintained so set leading pointer to 2 and trailing pointer to 1. If the leading pointer pointers to a number that is the same as the previous 2 then increment the lead pointer (skip the extra repeated number). Otherwise increment the trail pointer and update the value of the trailpointer to be equal to the value at the lead pointer.