Update 134. Gas Station
Only need single pass, if the total cost is higher than the total gas then no solution exists. Otherwise keep track of start position by resetting every time the current tank becomes negative (if the current tank becomes negative this implies that there is a surplus that is required to reach the next station that can't be achieved from the previous starting point.