Create 135. Candy
Greedy algorithm, initialize candy array with 1 to satisfy first requirement. Then going from left to right increment the current index if the current rating is higher than the rating of the child on the left (left comparison). Then going from right to left, increment the current index if the current rating is high than the rating of the child on the right, but need to make sure that the current value isnt already larger than the value on the right from the previous pass. Then sum candy array and return