Create 1758. Minimum Changes To Make Alternating Binary String
Explained in pattern, end result can either be 0101 or 1010 format. Can check if each odd digit is 0 and each even digit is 1 increasing a counter for each discrepancy. The alternate solution would require the total number of digits - the current count so taking the minimum of these two values will yield the minimum result.