Update 76. Minimum Window Substring
SLIDING WINDOW PROBLEM. Better solution, don't need messy comparison iteration. Just need to keep a counter of the number of "good" chars you have seen and then you know you have a good solution when the number of good chars matches the length of the string you are searching for. Important to know that you only increment or decrement the counter for good chars if the value is less than that in the original string (this handles duplicates because increasing count of letter above what is required wont increase count and decreasing count while remaining above required won't affect answer from being "correct"