Create 846. Hand of Straights
Can sort array and then keep track of how full each hand is. If the current working array is empty just add the next number, otherwise can return false if the difference between current number and last seen number in that array is more than 1 can return false. If the number is the same then another array needs to be created, if the number of arrays created exceeds the number of hands then can return false. Otherwise keep checking the next value in the array until all values are accounted for.

Also implemented quick sort for practice.